Moving out storage from project explorer
