+++ Squalr/Source/ProjectExplorer/ProjectExplorerViewModel.cs
@@ -1,26 +1,19 @@
 ﻿namespace Squalr.Source.ProjectExplorer
 {
-    using Controls;
     using Docking;
     using Engine;
     using Engine.OperatingSystems;
     using GalaSoft.MvvmLight.Command;
     using Main;
-    using Microsoft.Win32;
-    using Output;
     using ProjectItems;
     using PropertyViewer;
     using Squalr.Properties;
-    using Squalr.Source.Analytics;
     using System;
     using System.Collections.Generic;
     using System.Collections.ObjectModel;
-    using System.IO;
     using System.Linq;
-    using System.Runtime.Serialization.Json;
     using System.Threading;
     using System.Threading.Tasks;
-    using System.Windows;
     using System.Windows.Input;
     using Utils.Extensions;
 
@@ -83,13 +76,14 @@
         {
             this.ContentId = ProjectExplorerViewModel.ToolContentId;
             this.ObserverLock = new Object();
+            this.ProjectItemStorage = new ProjectItemStorage();
 
             // Commands to manipulate project items may not be async due to multi-threading issues when modifying collections
-            this.OpenProjectCommand = new RelayCommand(() => this.OpenProject(), () => true);
-            this.ImportProjectCommand = new RelayCommand(() => this.ImportProject(), () => true);
-            this.ExportProjectCommand = new RelayCommand(() => this.ExportProject(), () => true);
-            this.ImportSpecificProjectCommand = new RelayCommand<String>((filename) => this.ImportProject(false, filename), (filename) => true);
-            this.SaveProjectCommand = new RelayCommand(() => this.SaveProject(), () => true);
+            this.OpenProjectCommand = new RelayCommand(() => this.ProjectItemStorage.OpenProject(), () => true);
+            this.ImportProjectCommand = new RelayCommand(() => this.ProjectItemStorage.ImportProject(), () => true);
+            this.ExportProjectCommand = new RelayCommand(() => this.ProjectItemStorage.ExportProject(), () => true);
+            this.ImportSpecificProjectCommand = new RelayCommand<String>((filename) => this.ProjectItemStorage.ImportProject(false, filename), (filename) => true);
+            this.SaveProjectCommand = new RelayCommand(() => this.ProjectItemStorage.SaveProject(), () => true);
             this.SelectProjectItemCommand = new RelayCommand<ProjectItem>((projectItem) => PropertyViewerViewModel.GetInstance().SetTargetObjects(projectItem), (projectItem) => true);
             this.AddNewAddressItemCommand = new RelayCommand(() => this.AddNewPointerItem(), () => true);
             this.AddNewScriptItemCommand = new RelayCommand(() => this.AddNewScriptItem(), () => true);
@@ -157,38 +151,9 @@
         public ICommand ClearSelectionCommand { get; private set; }
 
         /// <summary>
-        /// Gets or sets a value indicating whether there are unsaved changes.
+        /// Gets an object that allows accesses to saving and loading project items.
         /// </summary>
-        public Boolean HasUnsavedChanges
-        {
-            get
-            {
-                return this.hasUnsavedChanges;
-            }
-
-            set
-            {
-                this.hasUnsavedChanges = value;
-                this.RaisePropertyChanged(nameof(this.HasUnsavedChanges));
-            }
-        }
-
-        /// <summary>
-        /// Gets the path to the project file.
-        /// </summary>
-        public String ProjectFilePath
-        {
-            get
-            {
-                return this.projectFilePath;
-            }
-
-            private set
-            {
-                this.projectFilePath = value;
-                this.RaisePropertyChanged(nameof(this.ProjectFilePath));
-            }
-        }
+        public ProjectItemStorage ProjectItemStorage { get; private set; }
 
         /// <summary>
         /// Gets the root that contains all project items.
@@ -200,7 +165,7 @@
                 return this.projectItems;
             }
 
-            private set
+            set
             {
                 this.projectItems = value;
                 this.RaisePropertyChanged(nameof(this.ProjectItems));
@@ -255,43 +220,6 @@
         public void DisableAllProjectItems()
         {
             this.ProjectItems?.ForEach(item => item.IsActivated = false);
-        }
-
-        /// <summary>
-        /// Prompts the user to save the project if there are unsaved changes.
-        /// </summary>
-        /// <returns>Returns false if canceled, otherwise true.</returns>
-        public Boolean PromptSave()
-        {
-            if (!this.HasUnsavedChanges)
-            {
-                return true;
-            }
-
-            String projectName = Path.GetFileName(this.ProjectFilePath);
-
-            if (String.IsNullOrWhiteSpace(projectName))
-            {
-                projectName = "Untitled";
-            }
-
-            MessageBoxResult result = CenteredDialogBox.Show(
-                "Save changes to project " + projectName + "?",
-                "Unsaved Changes",
-                MessageBoxButton.YesNoCancel,
-                MessageBoxImage.Question);
-
-            if (result == MessageBoxResult.Cancel)
-            {
-                return false;
-            }
-
-            if (result == MessageBoxResult.Yes)
-            {
-                this.SaveProject();
-            }
-
-            return true;
         }
 
         /// <summary>
@@ -417,313 +345,6 @@
                     Thread.Sleep(SettingsViewModel.GetInstance().TableReadInterval);
                 }
             });
-        }
-
-        /// <summary>
-        /// Opens a project from disk.
-        /// </summary>
-        private void OpenProject()
-        {
-            OpenFileDialog openFileDialog = new OpenFileDialog();
-            openFileDialog.Filter = ProjectExtensionFilter;
-            openFileDialog.Title = "Open Project";
-
-            if (openFileDialog.ShowDialog() == false)
-            {
-                return;
-            }
-
-            this.ProjectItems = new ObservableCollection<ProjectItem>();
-            this.ProjectFilePath = openFileDialog.FileName;
-
-            // Open the project file
-            try
-            {
-                if (!File.Exists(this.ProjectFilePath))
-                {
-                    OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Error, "Unable to locate project.");
-                    return;
-                }
-
-                using (FileStream fileStream = new FileStream(this.ProjectFilePath, FileMode.Open, FileAccess.Read))
-                {
-                    DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(ProjectItem[]));
-                    this.ProjectItems = new ObservableCollection<ProjectItem>(serializer.ReadObject(fileStream) as ProjectItem[]);
-                    this.HasUnsavedChanges = false;
-                }
-            }
-            catch (Exception ex)
-            {
-                this.ProjectFilePath = String.Empty;
-                OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Error, "Unable to open project", ex.ToString());
-                return;
-            }
-
-            // Open the hotkey file
-            try
-            {
-                String hotkeyFilePath = this.GetHotkeyFilePathFromProjectFilePath(this.projectFilePath);
-
-                if (File.Exists(hotkeyFilePath))
-                {
-                    using (FileStream fileStream = new FileStream(hotkeyFilePath, FileMode.Open, FileAccess.Read))
-                    {
-                        DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(ProjectItemHotkey[]));
-                        ProjectItemHotkey[] projectItemHotkeys = serializer.ReadObject(fileStream) as ProjectItemHotkey[];
-
-                        this.BindHotkeys(this.ProjectItems, projectItemHotkeys);
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Warn, "Unable to open hotkey profile", ex);
-                AnalyticsService.GetInstance().SendEvent(AnalyticsService.AnalyticsAction.General, ex);
-                return;
-            }
-        }
-
-        /// <summary>
-        /// Imports a project from disk, adding the project items to the current project.
-        /// </summary>
-        /// <param name="filename">The file path of the project to import.</param>
-        private void ImportProject(Boolean resetGuids = true, String filename = null)
-        {
-            // Ask for a specific file if one was not explicitly provided
-            if (filename == null || filename == String.Empty)
-            {
-                OpenFileDialog openFileDialog = new OpenFileDialog();
-                openFileDialog.Filter = ProjectExtensionFilter;
-                openFileDialog.Title = "Import Project";
-
-                if (openFileDialog.ShowDialog() == false)
-                {
-                    return;
-                }
-
-                filename = openFileDialog.FileName;
-
-                // Clear the current project, such that on save the user is prompted to reselect this
-                this.ProjectFilePath = null;
-            }
-
-            // Import the project file
-            ProjectItem[] importedProjectRoot = null;
-
-            try
-            {
-                if (!File.Exists(filename))
-                {
-                    OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Error, "Unable to locate project.");
-                    return;
-                }
-
-                using (FileStream fileStream = new FileStream(filename, FileMode.Open, FileAccess.Read))
-                {
-                    DataContractJsonSerializer deserializer = new DataContractJsonSerializer(typeof(ProjectItem[]));
-                    importedProjectRoot = deserializer.ReadObject(fileStream) as ProjectItem[];
-
-                    // Add each high level child in the project root to this project
-                    foreach (ProjectItem child in importedProjectRoot)
-                    {
-                        this.AddNewProjectItems(false, child);
-                    }
-
-                    this.HasUnsavedChanges = true;
-                }
-            }
-            catch (Exception ex)
-            {
-                OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Error, "Unable to import project", ex);
-                AnalyticsService.GetInstance().SendEvent(AnalyticsService.AnalyticsAction.General, ex);
-                return;
-            }
-
-            // Import the hotkey file
-            try
-            {
-                String hotkeyFilePath = this.GetHotkeyFilePathFromProjectFilePath(filename);
-
-                if (File.Exists(hotkeyFilePath))
-                {
-                    using (FileStream fileStream = new FileStream(hotkeyFilePath, FileMode.Open, FileAccess.Read))
-                    {
-                        DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(ProjectItemHotkey[]));
-                        ProjectItemHotkey[] projectItemHotkeys = serializer.ReadObject(fileStream) as ProjectItemHotkey[];
-
-                        // Bind the hotkey to this project item
-                        this.BindHotkeys(importedProjectRoot, projectItemHotkeys);
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Warn, "Unable to open hotkey profile", ex);
-                AnalyticsService.GetInstance().SendEvent(AnalyticsService.AnalyticsAction.General, ex);
-                return;
-            }
-
-            // Randomize the guid for imported project items, preventing possible conflicts
-            if (resetGuids && importedProjectRoot != null)
-            {
-                foreach (ProjectItem child in importedProjectRoot)
-                {
-                    child.ResetGuid();
-                }
-            }
-        }
-
-        /// <summary>
-        /// Save a project to disk.
-        /// </summary>
-        private void SaveProject()
-        {
-            // Save the project file
-            try
-            {
-
-                SaveFileDialog saveFileDialog = new SaveFileDialog();
-                saveFileDialog.Filter = ProjectExplorerViewModel.ProjectExtensionFilter;
-                saveFileDialog.Title = "Save Project";
-                saveFileDialog.FileName = String.IsNullOrWhiteSpace(this.ProjectFilePath) ? String.Empty : Path.GetFileName(this.ProjectFilePath);
-                saveFileDialog.RestoreDirectory = true;
-                saveFileDialog.InitialDirectory = String.IsNullOrWhiteSpace(this.ProjectFilePath) ? String.Empty : Path.GetDirectoryName(this.ProjectFilePath);
-
-                if (saveFileDialog.ShowDialog() == true)
-                {
-                    this.ProjectFilePath = saveFileDialog.FileName;
-
-                    using (FileStream fileStream = new FileStream(this.ProjectFilePath, FileMode.Create, FileAccess.Write))
-                    {
-                        DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(ProjectItem[]));
-                        serializer.WriteObject(fileStream, this.ProjectItems);
-                    }
-
-                    this.HasUnsavedChanges = false;
-                }
-            }
-            catch (Exception ex)
-            {
-                OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Fatal, "Unable to save project", ex);
-                AnalyticsService.GetInstance().SendEvent(AnalyticsService.AnalyticsAction.General, ex);
-                return;
-            }
-
-            // Save the hotkey profile
-            try
-            {
-                String hotkeyFilePath = this.GetHotkeyFilePathFromProjectFilePath(this.projectFilePath);
-
-                using (FileStream fileStream = new FileStream(hotkeyFilePath, FileMode.Create, FileAccess.Write))
-                {
-                    DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(ProjectItemHotkey[]));
-                    ProjectItemHotkey[] hotkeys = ProjectExplorerViewModel.GetInstance().ProjectItems?.Select(x => new ProjectItemHotkey(x.HotKey, x.Guid)).ToArray();
-                    serializer.WriteObject(fileStream, hotkeys);
-                }
-            }
-            catch (Exception ex)
-            {
-                OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Error, "Unable to save hotkey profile", ex);
-                AnalyticsService.GetInstance().SendEvent(AnalyticsService.AnalyticsAction.General, ex);
-                return;
-            }
-        }
-
-        /// <summary>
-        /// Export a project to separate files.
-        /// </summary>
-        private void ExportProject()
-        {
-            Task.Run(() =>
-            {
-                // Export the project items to thier own individual files
-                try
-                {
-                    if (String.IsNullOrEmpty(this.ProjectFilePath) || !Directory.Exists(Path.GetDirectoryName(this.ProjectFilePath)))
-                    {
-                        OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Info, "Please save the project before exporting");
-                        return;
-                    }
-
-                    OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Info, "Project export starting");
-
-                    String folderPath = Path.Combine(Path.GetDirectoryName(this.ProjectFilePath), "Export");
-                    Directory.CreateDirectory(folderPath);
-
-                    Parallel.ForEach(
-                        this.ProjectItems,
-                        SettingsViewModel.GetInstance().ParallelSettingsFast,
-                        (projectItem) =>
-                    {
-                        ProjectItem targetProjectItem = projectItem;
-
-                        if (projectItem is ScriptItem)
-                        {
-                            ScriptItem scriptItem = projectItem as ScriptItem;
-
-                            if (!scriptItem.IsCompiled)
-                            {
-                                targetProjectItem = scriptItem?.Compile();
-                            }
-                        }
-
-                        if (targetProjectItem == null)
-                        {
-                            return;
-                        }
-
-                        String filePath = Path.Combine(folderPath, targetProjectItem.Description + ProjectExplorerViewModel.ProjectFileExtension);
-
-                        using (FileStream fileStream = new FileStream(filePath, FileMode.Create, FileAccess.Write))
-                        {
-                            List<ProjectItem> newProjectRoot = new List<ProjectItem>();
-                            newProjectRoot.Add(targetProjectItem);
-
-                            DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(ProjectItem[]));
-                            serializer.WriteObject(fileStream, newProjectRoot.ToArray());
-                        }
-                    });
-                }
-                catch (Exception ex)
-                {
-                    OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Fatal, "Unable to complete export project", ex);
-                    AnalyticsService.GetInstance().SendEvent(AnalyticsService.AnalyticsAction.General, ex);
-                    return;
-                }
-
-                OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Info, "Project export complete");
-            });
-        }
-
-        /// <summary>
-        /// Binds deserailized project item hotkeys to the corresponding project items.
-        /// </summary>
-        /// <param name="projectItems">The candidate project items to which we are binding the hotkeys.</param>
-        /// <param name="projectItemHotkeys">The deserialized project item hotkeys.</param>
-        private void BindHotkeys(IEnumerable<ProjectItem> projectItems, IEnumerable<ProjectItemHotkey> projectItemHotkeys)
-        {
-            if (projectItemHotkeys.IsNullOrEmpty())
-            {
-                return;
-            }
-
-            projectItemHotkeys
-                .Join(
-                    projectItems,
-                    binding => binding.ProjectItemGuid,
-                    item => item.Guid,
-                    (binding, item) => new { binding = binding, item = item })
-                .ForEach(x => x.item.LoadHotkey(x.binding.Hotkey));
-        }
-
-        /// <summary>
-        /// Gets the hotkey file path that corresponds to this project file path.
-        /// </summary>
-        /// <param name="projectFilePath">The path to the project file.</param>
-        /// <returns>The file path to the hotkey file.</returns>
-        private String GetHotkeyFilePathFromProjectFilePath(String projectFilePath)
-        {
-            return Path.Combine(Path.GetDirectoryName(projectFilePath), Path.GetFileNameWithoutExtension(projectFilePath)) + ProjectExplorerViewModel.HotkeyFileExtension;
         }
     }
     //// End class
