+++ Squalr/Source/ProjectExplorer/ProjectItems/AddressItem.cs
@@ -102,7 +102,7 @@
                 // Clear our current address value
                 this.addressValue = null;
 
-                ProjectExplorerViewModel.GetInstance().HasUnsavedChanges = true;
+                ProjectExplorerViewModel.GetInstance().ProjectItemStorage.HasUnsavedChanges = true;
                 this.NotifyPropertyChanged(nameof(this.DataType));
             }
         }
@@ -148,7 +148,7 @@
                 }
 
                 this.isValueHex = value;
-                ProjectExplorerViewModel.GetInstance().HasUnsavedChanges = true;
+                ProjectExplorerViewModel.GetInstance().ProjectItemStorage.HasUnsavedChanges = true;
                 this.NotifyPropertyChanged(nameof(this.IsValueHex));
             }
         }
